Updated update.js
Simplified the code in remoteHack, removing a duplicate section with better IF management. Also added a selection for directing remote hack servers to grow and weaken their targets, rather than hack them, when the server list is short. This should allow the botlist to focus on using their limited threads to the benefit of the much larger botnet & home processes, while not getting in the way with an inopportune hack.